+++ LeetCode/SubarraySumEqualsK.cpp
@@ -11,7 +11,7 @@ public:
         unordered_map<int, int> m;
         m[0] = 1;
         int sum = 0;
-        for (int i=0; i<nums.size(); i++)
+        for (int i=0; i<(int)nums.size(); i++)
         {
             sum += nums[i];
             if (m.find(sum - k) != m.end())
@@ -23,11 +23,11 @@ public:
     
     int subarraySumSlow(vector<int>& nums, int k) {
         int count = 0;
-        for (int i=0; i<nums.size(); i++)
+        for (int i=0; i<(int)nums.size(); i++)
         {
             int sum = nums[i];
             if (sum == k) count++;
-            for (int j=i+1; j<nums.size(); j++)
+            for (int j=i+1; j<(int)nums.size(); j++)
             {
                 sum += nums[j];
                 if (sum == k) count++;
